Move `ReplyBarConfiguration` child types to new file
