+++ Mlem/App/Enums/Interaction/ReplyBarConfiguration+Types.swift
@@ -1,0 +1,111 @@
+//
+//  ReplyBarConfiguration+Types.swift
+//  Mlem
+//
+//  Created by Sjmarf on 2026-03-15.
+//
+
+import Actions
+import Foundation
+import MlemMiddleware
+import SwiftUI
+
+extension ReplyBarConfiguration {
+    enum ActionType: String, ActionTypeProviding {
+        typealias Configuration = ReplyBarConfiguration // swiftlint:disable:this nesting
+        
+        case upvote
+        case downvote
+        case save
+        case reply
+        case markRead
+        case selectText
+        case report
+        
+        static var defaultWidgets: [ActionType] { [
+            .upvote,
+            .downvote,
+            .save,
+            .reply,
+            .markRead
+        ] }
+        
+        var appearance: ActionAppearance {
+            switch self {
+            case .upvote: .upvote(isOn: false)
+            case .downvote: .downvote(isOn: false)
+            case .save: .save(isOn: false)
+            case .reply: .reply()
+            case .markRead: .markRead(isOn: false)
+            case .selectText: .selectText()
+            case .report: .report()
+            }
+        }
+        
+        func associatedReadouts(context: any InteractableProviding) -> Set<ReplyBarConfiguration.ReadoutType> {
+            switch self {
+            case .upvote: context.votes.value?.myVote ?? .none == .upvote ? [.upvote, .score] : [.upvote]
+            case .downvote: context.votes.value?.myVote ?? .none == .downvote ? [.downvote, .score] : [.downvote]
+            case .save: [.saved]
+            case .reply, .markRead, .selectText, .report: []
+            }
+        }
+    }
+    
+    enum CounterType: String, CounterTypeProviding {
+        typealias Configuration = ReplyBarConfiguration // swiftlint:disable:this nesting
+        
+        case score
+        case upvote
+        case downvote
+        case reply
+        
+        static var defaultWidgets: [CounterType] { allCases }
+        
+        var appearance: CounterAppearance {
+            switch self {
+            case .score: .score()
+            case .upvote: .upvote()
+            case .downvote: .downvote()
+            case .reply: .reply()
+            }
+        }
+        
+        func associatedReadouts(context: any InteractableProviding) -> Set<ReplyBarConfiguration.ReadoutType> {
+            switch self {
+            case .score: [.upvote, .downvote, .score]
+            case .upvote: context.votes.value?.myVote ?? .none == .upvote ? [.upvote, .score] : [.upvote]
+            case .downvote: context.votes.value?.myVote ?? .none == .downvote ? [.downvote, .score] : [.downvote]
+            case .reply: []
+            }
+        }
+    }
+    
+    enum ReadoutType: String, ReadoutTypeProviding {
+        case created
+        case score
+        case upvote
+        case downvote
+        case comment
+        case saved
+        
+        var appearance: MockReadoutAppearance {
+            switch self {
+            case .created: .init(icon: .general.time, label: "18h")
+            case .score: .init(icon: .lemmy.votes, label: "7")
+            case .upvote: .init(icon: .lemmy.upvoted, label: "9")
+            case .downvote: .init(icon: .lemmy.downvoted, label: "2")
+            case .comment: .init(icon: .lemmy.replies, label: "1")
+            case .saved: .init(icon: .lemmy.saved, label: "")
+            }
+        }
+        
+        func compatibleWith(otherReadouts: Set<Self>) -> Bool {
+            switch self {
+            case .score: otherReadouts.isDisjoint(with: [.upvote, .downvote])
+            case .upvote, .downvote: !otherReadouts.contains(.score)
+            default: true
+            }
+        }
+    }
+}
